feat: configure backend to serve frontend
- Update backend endpoints to match frontend requirements
- Configure CORS settings for cross-origin communication
- Add static directory for serving images
- Enable frontend to access backend resources and images

+++ backend/src/controllers/products.controller.js
@@ -95,8 +95,9 @@ module.exports = {
         });
       }
       // remove image
-      fs.unlinkSync(`./src/uploads/${product[0].linkImage}`);
-
+      if (fs.existsSync(`./src/uploads/${product[0].linkImage}`)) {
+        fs.unlinkSync(`./src/uploads/${product[0].linkImage}`);
+      }
       await productService.deleteProduct(productID);
 
       res.status(200).json({
